feat(arch): module-level import rules within a layer (#1849, #1843)
The #668 blocks fix the DIRECTION between layers. Inside a layer they have
nothing to say, and a wrong import is invisible until someone reads the file —
which is how `approval.ts` came to import `./conversation` to build a UI
string. Three edges the codebase already documents in prose, now frozen:

- `llm/approval.ts` ↛ `llm/conversation` — the approval engine owns tier policy
  and the proposal lifecycle, not conversation storage.
- `history/**` ↛ `notebase/**` — history sits BELOW notebase (`notebase/fs`
  calls its capture hooks; restore orchestration composes both in the IPC
  layer). The reverse is the cycle that arrangement exists to avoid.
- `graph/**` ↛ `llm/**` — the graph is the substrate, the LLM is a client of
  it, and the approval engine that would want the edge already depends on
  graph.

Deliberately a short list. It isn't an attempt to describe the design — only
the edges whose absence something else depends on.

Making the first rule true meant doing #1843: `proposalCause` moves to
`llm/proposal-cause.ts`, so the lifecycle asks a named module for a label
instead of knowing how a label is found. Behaviour is unchanged and its tests
pass untouched.

Each rule was verified by writing the import it forbids and watching it fail.
That caught a real defect in my first draft: `no-restricted-imports` matches
the specifier as WRITTEN, not the resolved path, so `**/llm/conversation`
silently never fired against `'./conversation'`. A rule that can't fire is
worse than no rule — it reads as protection. The glob now covers both forms,
and the reason is recorded next to it.
